Refactor Android codebase
