+++ android/src/main/kotlin/com/kasem/media_picker_builder/model/MediaFile.kt
@@ -1,4 +1,4 @@
-package com.kasem.media_picker_builder
+package com.kasem.media_picker_builder.model
 
 import org.json.JSONObject
 
@@ -49,10 +49,7 @@ class MediaFile(
         return id.hashCode()
     }
 
-
     enum class MediaType {
         IMAGE, VIDEO
     }
-
-
 }
